Changes to sorting; $USER value instead of <default>
Before:
- sort everything by domain first, user next

Now:
- list domains in the order of discovery
- sort by user per domain

Sorting by domain is not a trivial task. If sorting as a whole, it would put
n1.com ahead of n.com, because it compares "1" to "." at position 2. If sorting
by parts, which side should you start comparing from AND how do you compare
domains with different number of parts?

Let it be the user who decides on the order.